oop-scripts/pyproject.toml
[tool.poetry]
name = "oop-scripts"
version = "0.1.0"
description = "ai tools for create oop practice lists"
authors = ["819110812@qq.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<4.0,>=3.9"
langchain = "^0.3.4"
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.2.3"
langchain-openai = "^0.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
